requirements.txt
admin-totals==1.0.1
adyengo==0.0.10
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.9
django-3-jet==1.0.7
django-admin-autocomplete-filter==0.6.1
django-admin-filter==0.1
django-admin-rangefilter==0.6.3
django-author==1.0.2
django-braces==1.13.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-csp==3.7
django-filter==2.4.0
django-heroku==0.3.1
django-jet==1.0.8
django-nucleus==0.1.2
django-object-actions==3.0.1
django-paypal==1.0.0
gunicorn==20.0.4
idna==2.10
ipaddress==1.0.23
jdatetime==3.6.2
jsonfield==3.1.0
persiantools==1.5.1
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
setuptools-git==1.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
